Fix shrinkwrap URLs for Git sources
This is a far-from-ideal fix, because the shrink-wrapped URL for any
bundle that relies on Git sources isn't *technically* shrinkwrapped –
the HEAD of that source will be loaded.

So we're fixing the issue where an incorrect Git source is used for some
shrink-wrapped bundles, but we've uncovered a new issue: shrink-wrapped
bundles aren't always shrink-wrapped :/

I'll open a separate issue for that.

This fixes #26.