Fix base URL calculation in demos (#141)
This turned out to be a bit of a nightmare. I've fixed it but it isn't
that pretty and this whole part of the codebase could do with a review
later. For now I'd quite like to get this into QA for testing and then
live to fix the demo issue.