Only allow origami components in demos
Update the demos controller to perform two validation steps.

1. Validate whether the components are from the FT bower registry or the global registry (network request).
2. Validate whether the components have an origami manifest post-install

Errors in either scenario result in a 400 error with an appropriate error messages.

There are currently 4 failing integration tests which attempt to build demos for non-existent modules.
These obviously can't be found in the FT bower registry, so now result in 400s rather than 560 errors.
As such, they're candidates for removal.