Prevent race condition when building branded CSS. (#192)
* Prevent race condition when building branded CSS.

If two CSS bundles are building at the same time, and they are identical except for the brand, there is a race condition which
means the same bundle will be returned for both. The impact is that the built CSS for one of the requests is for the incorrect
brand.

This happens because the build service writes a Sass file to build which includes the brand Sass variable. The Sass filename was
unique for each module set, but not for each module set and brand.