Copy files from `Financial-Times/obt5fork`
These will eventually replace the dependency on
`origami-build-tools` v5.

Tests have been copied under `unit` although they are more
functional tests as they interact with the filesystem.

`obtv5fork` is a fork of `origami-build-tools` v5, which the Origami
Build Service currently depends on. `obtv5fork` removes features
which are for local component development and not needed by the
Origami Build Service.

The original intention was to replace the dependency on
`origami-build-tools` with `obtv5fork` but the team decided
to move the codebase within Origami Build Tools instead.

For a history of these files see:
https://github.com/Financial-Times/obt5fork