Move `obt5fork` tests to the `integration` test directory.
They're currently failing due to `Mockery` which is used
within unit tests.

These tests are more like functional tests. They test an isolated
part of the codebase but interacting with the dependencies and the
filesystem.

This is part of the work to replace origami-build-tools
with code copied from https://github.com/Financial-Times/obt5fork/