Add middleware to ensure all requested Origami modules are built using the same Origami specification version
This uses a fixed list (json file) of Origami Components to check if something is an Origami Component.

If the requested module is an Origami Component and is using a Git reference instead of a SemVer version, we assume they are requesting a version of the Origami Component which was built using Origami Spec v1.