Do not crash the app if the NPM cache fails to prime.
This might happen if a component is not published for some reason.
The Build Service failed recently because of this:
https://financialtimes.slack.com/archives/C02FU5ARJ/p1607506789255200
`@financial-times/o-fonts-assets` was published to the test npm
registry but not to the official npm registry.

When v2 components are on the official npm registry we can be
confident that our tests will catch a future regression like this.
Supports: https://github.com/Financial-Times/origami/pull/86