Validate system-codes with the new http api
This changes how we validate the system_code query parameter, previously we were checking it adhered to the same regular expression that is currently used within the Biz-Ops project. This approach was not complete because it never proved that the provided system_code value was in fact a real system code stored within Biz-Ops, which is ideally what we want to do.

There is a new HTTP API available to validate whether something is a real system code within Biz-Ops, https://system-codes.in.ft.com/ - This is what we are now using instead of the regular expression.

The HTTP API is implemented entirely within Fastly VCL, which means it is fast to respond and very unlikely to go offline. If it does take a long time to respond or does infact go offline we will 'fail open' and pretend that the provided system_code value is a real valid system code. This ensures that we don't break customers websites simpliy because we can't validate their provided system_code.