Install `o-fonts-assets` with Github for the v3 files endpoint
We install from the Github repository because `o-fonts-assets` is
not available via npm as it must remain private for license reasons.

With our current Github license we could publish `o-fonts-assets`
to the Github packages registry privately. However we would have
to use a different scope to `@financial-times` so that we can install
from both the Github and npm registry. Unless we were to move all
`@financial-times` packages to the Github packages registry. We
think the later would be the best approach but would take time
to propose and implement across the ft. Hence for now we opted
to install the Github repository directly.