Add URL Builder
The URL Builder accepts an existing Build Service URL and builds
an updated URL which includes the latest version number for all
components, helping users to upgrade their Build Service requests
and migrate to the new version of components. To further help a
table is displayed which shows the components which were requested,
the latest version number for the component, a label to show if it
is up to date or not, and a link to the readme to access the
migration guides.

So far it does not handle all cases, e.g:
- Only the "bundles" endpoint is supported
- Module filepaths and non-Origami components aren't supported
- If the latest version of components can only be used with the
(upcoming) v3 Build Service endpoint then the generated URL will
include a v3 Build Service url but will not warn about any
parameters used which are  no longer be supported in v3 of the
Build Service