Restrict demo builds to valid semver releases (v2)
v2 will only build demos for components registered with the Origami
Bower Registry, all from an allow list of Github organisations.
Hower it is possible to fork a component on Github and open a
pull request with malicious code changes and use the git ref that
creates to build a malicious demo through the Build Service, hosted
on the ft.com domain. By restricting demo builds to valid semver
releases this is prevented.