Implement memoization for the v3 api routes with a 2 minute cache
This commit makes all the v3 endpoints have an in-memory cache which stores the results for 2 minutes.

The reason to add this cache is because sometimes a response will take more than 30 seconds to generate (usually because the `npm install` is taking a while to finish) and we Heroku's router will close any connections which have not returned a response within 30 seconds.

In a future commit we can add logic to detect when a response is going to take longer than 30 seconds and we can respond with a redirect back to the same endpoint with the same parameters. The browser will follow the redirect and the server will call the memoized function with the same parameters as before, which will return a Promise for the response already being generated based on the previous request instead of generating a new response from scratch.