Move CDN specific caching logic to the surrogate-control header
The surrogate-control header is only read by the Fastly CDN and not by other user-agents such as a web browser.

Fastly will remove the surrogate-control header from the response before they forward it to the user-agent, which is a very useful feature to have when the Fastly setup uses Shielding, which our Fastly setup uses. The reason it is useful is because it solves a problem which occurs when a non-shield node makes a request to a shield node and the shield-node has a stale response it is allowed to use. If we don't use the surrogate-control header, and instead use cdn caching directives available in the cache-control header, Fastly will send the cache-control header unmodified from the shield-node to the non-shield-node, this would include the cdn specific caching directives, which is not what we want because the content is stale but the non-shield-node is now caching it as non-stale content (caching it for a longer time than we want). If we use the surrogate-control header for the cdn specific caching rules, then in the previous scenario, the shield-node would remove the surrogate-control header before sending the response to the non-shield-node and the non-shield-node would only cache the content for the same amount of time that any other user-agent would