Fix OBS v2 endpoints given component dependencies with invalid origami.json
`/v2/files/o-header@2.5.13/img/ft-logo.png` currently fails as an install
of `o-viewport`, a dependency, has an invalid `origami.json` that cannot
be parsed.

Instead, when it is not possible to check if a component follows the v1 spec
due to `origami.json` issues assume that it does so not to break existing
requests.

Related PR: https://github.com/Financial-Times/origami-build-service/pull/438
Fixes: https://sentry.io/organizations/financial-times/issues/2254274035/?project=52558&query=is%3Aunresolved