Escape `script` tags in compiled demo JS.
Demo JS is inlined in demo HTML. When a component's JS included
a `script` tag that was inlined within the demo HTML the component
`script` tag closed the `script` tag in the demo html early. This
meant demo JS did not run and was visible as text in the demo.
To avoid that escape any closing script tag `</script>` before
injecting it into demo HTML.

This was an issue for `o-autocomplete@1.0.0`.