bugfix: make the v3 exported Origami variable to not have the `.default` property on the components
We want to match the behavior of the v2 export Origami variable which is to have the components' code be available directly from the variable. E.G.`let Table = window.Origami['o-table'];` should work in v2 and in v3.

Currently it does not do this due to changing the bundler being used to ESBuild, which puts the exports component code on a `.default` property.

This is a breaking change, but we should try and do it if possible because right now it complicates the migration process more than we want and is behavior we would like to not have.

We should check the logs to see if anyone is using v3 in production and see if this behavior is already being relied upon. I know origami services moved to the v3 api but they do not rely on the exported Origami variable, hopefully this is the same for anything else which is using v3 right now 🤞