Update test expectactions.
The existing integration test for the `default` property regression
failed as `document` is not avaliable in the test enviroment. It's
referenced in most component's `main.js` file to listen for the
`o.DOMContentLoaded` event.

Instead test with `o-test-component` which does not do this. As
`o-autoinit` previously threw an error if `.default` was assigned
to also confirm the generated bundle for `o-autoinit` does not
assign the `default` property.