configure npm to use a cache directory which is stored within the app this enables us to have commit cached files into the project and have them always be in the npm cache for the build-service to use
when installing dependencie - install in offline mode first if possible then fallback to network if required.
this should allow the service to have the best possible chance to have a fast install as `npm install --offline` skips lots of work if all the packages are in the cache. If a package is not in the cache then npm will error and we will run `npm install --prefer-offline` which allows npm to make network requests to fill in the missings parts of the cache.

The reason we don't run `npm install --prefer-offline` all the time is because npm will make a bunch of network requests even if all the packages are in the cache, which is much slower than running `npm install --offline` first.