feat: Implement archive for Origami Build Service v2.
Includes an env to turn the future on/off/fallback to existing behaviour.

Logs requests which miss the archive to Splunk when set to fallback to
current behaviour.

When using the archive hits/miss count is logged to grafana.

Handles bundle, demo, and file routes.
Does not handle module route (component meta data). This isn't used anymore.

Adds a script to populate archive. Will run with more urls soon.

May move v1 archive to use the same mecanism in the near future,
or delete if they are no londer relied on.